Code refactor for module tests
Desired feature is to pass list of commands to mumjolandia to
automatically execute them and then exit. It will allow to easily
introduce module tests.

Added features:
- CommandFactory can take list as an argument and returns list of commands
- Console has async getCommand function (it return None until command is
passed, so main thread is not blocked waiting for input
- Main loop of MumjolandiaCli is refactored and can execute passed list of
commands
- MumjolandiaStarter accepts list of strings to be translated to
commands and passed to MumjolandiaCli
- Unit tests for MumjolandiaCli after refactor can be a lot more simple
and does not need to use workarounds. Those workarounds worked only for
older code, so part of them are adjusted to work with new loop, but some
ut-s are failing now