Task 'finish date' management
It is desired to add task without specified date to finish it (now it
was added for 'today') . This change makes all added tasks to be
unassigned to any day. Now it is required to add finish date by 'set'
command. It is also possible to list all unassigned tasks by 'list x'
command, and clear 'finish date' of task by 'set none' command.