Changing games design
Now games are handled by mysql database. A whole process of creating
and managing of database is introduced. It is now possible to list,
add and remove games to/from database. Game by far has only a title,
but this change will allow to introduce desired functionality in the
future.