Add password module
The purpose of this module is to store in memory encrypted usernames and
passwords for services. They are enxrypted using external DES
implementation and decoded/encoded using provided user password.

Main implemented supervisor commands and way of working:

1. Command 'init' - set main password (preferably only once)
2. Command 'set' - enter main password (every time after running
   mumjolandia)
3. Commands 'list', 'remove', 'add', 'get' - self explanatory

Even if this is perfectly fine to decode passwords with incorrect main
password (and receiving trash data), supervisor uses magic word to
determine if password entered with 'set' command is correct, to avoid
ie. adding new password with different key and producing mess in data.

Passwords are stored via picke file and main password's validation
string is sotored in shared preferences.