Refactor commands shorcuts
There are handy shortucts for commands ie. 'rm' instead of 'remove' that
were parsed by function in cli thread. This solution was inconsistant
and produced confusion. As it is possible to add synonims to supervisor
dictionaries and link them to the save command implementation code is
changed to follow this solution.