Add 'exit' command at the end of parameters
It is possible to not pass 'exit' command at the end of mumjolandia
arguments. This use case has no use by now so 'exit' command will aways
be appended at the end of passed arguments list.
Cli is fixed to break out of parsing passed commands if 'exit' command
is exexuted (so exit flag is set up).