Add core of task reminder
Reminding tasks aka showing tasks for a couple of days before real
date_to_finish and providing funcitonality to mark those tasks as 'done'
is desired feature that was not possible to implement without serious
refactor and a lot of new code. Main reason of this are periodic tasks
which are not stored as "real" tasks, so their state can not be easily
manipulated (every periodic task is constantly repeated every few days).

As this feature is taking a lot of time to implement it will be
delivered by parts. This change provides functional reminding of
periodic tasks. To achieve this the following changes are introduced:
- Extracted creation of periodic tasks to PeriodicTaskLoader
- Periodic tasks are stored in 'PeriodicTask' objects, so they are no
longer translated 'in flight' to standard tasks. PeriodicTasks object
provides methods to determine next occurrence and information about
necessity of reminding in given day.
- Added 'reminder' to standard 'Task' object and 'TaskReminder' that
can be used to say if task should be reminded (currently functional
only with a debugger)
- Added DateHelper to provide global 'get_today' function that is easy
to mock. Changed uts and code to use it (not done fully)
- Added PeriodicTaskGenerator that returns periodic tasks for a given
day and next occurences of every periodic task.
- Adjusted PeriodicTaskProgressHandler to store and track state of
periodic tasks for time periods
- Added tests for TaskReminder, PeriodicTask and PeriodicTaskGenerator