Tasks are reminded for a whole day
Previously setting task with reminder provided exact datetime (ie. hour
and minute as well), so when task was added at 5pm and had to be
reminded in the next day, it wouldn't be shown until 5pm. After this
change this functionality is still working this way, but setting
reminder for a task procides date with hour and minutes zeroed, so they
are shown correctly.