Fix periodic tasks perofrmance and ut
Tasks are parsed once during object creation, so it is now a lot faster.
Downside of this approach is need of restarting app after changing xml
files with events and tasks. This can be adjusted to check automatically
if file has changed, but let's be real; i will never do this.
UTs are fixed as well.